backend/inference/main.py
import logging
import os

from dotenv import load_dotenv
from fastapi import FastAPI

load_dotenv()

logging.basicConfig(
    level=logging.INFO,
    format="%(asctime)s %(levelname)s %(name)s: %(message)s",
)

from routes.analyze import router as analyze_router
from routes.summary import router as summary_router

app = FastAPI(title="Voxidria Inference Service")
app.include_router(analyze_router)
app.include_router(summary_router)


@app.get("/health")
def health():
    return {"ok": True}


if __name__ == "__main__":
    import uvicorn
    uvicorn.run("main:app", host="0.0.0.0", port=int(os.getenv("PORT", 8000)), reload=False)
