backend/inference/requirements.txt
fastapi
uvicorn[standard]
requests
supabase
python-dotenv
numpy
scikit-learn
tensorflow
joblib
librosa
praat-parselmouth
pydub
